Use encoder classes for output size
Previously the number of classes in the train dataset was used as
the output size. That is to say that if there were 20 classes in
the training set, the output size of the model would be 20. As a
result, when the classes were weighted based on all labeled data,
the weight vector was a different length than the output in some
cases